backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
Werkzeug==3.0.1
